Added Average Power Load logging
Power load is averaged over 30 samples (1 minute) and logged to file.
Log file is created based on date, e.g. 20100129.log for the 29th of
January 2010.  The logged value is a percentage of UPS power rating.

The logging option is enabled/disabled through an entry in the tray
icon's context menu.